Improved the installation handling
This add new perspective on how to install a project without set-up the
properties to set the different install directories properties. This is
really useful when we create a project like a shared library that will be
used by a bigger projects where those projects will set-up those
properties in order to control finely the installation steps.

Please note that the install properties set at the root of the project
have always a high priority.

Finally, there is support for the command-lines in order to modify the
install properties - like the product name - by using the following
command:

$ qbs build -d ..\build-project StoiridhUtils.Project.productName:Project