Update Python 3 API reference
An API section has been added. Only the Python API is present and later
the Qbs API will be added.

Also note there are several improvements on the directives.  Finally,
the global.rst file is removed since there is no real use for now since
the glossary is already in the master-toc.