Update documentation (#11)
* Modify the indentation (3) for the reST files

The previous indentation was 4. The change follows the Python's style
guide with an exception:

The maximum line length is 100 characters instead of 80.

* Update version from the config file

Indeed, the actual version was 0.1.0, but the next release will be the
version 1.1.0.

* Update the references of the VersionNumber class

VersionNumber is into the stoiridh.qbs.tools package.

* Add a release notes section for the legacy versions

In the documentation, a Release Notes section has been added in order to
centralise the release notes of the legacy versions as well as the
future versions.

* Add a glossary

The glossary will contain all terms used in the documentation. Note also
there is an entry in the first page of the documentation in order to
find them all more easily.

* Update headers

In order to append my e-mail address.

* Add new epilogues

Those epilogues are available for all reST files of the documentation.

* Update the index (master) documentation

In order to use the epilogues and put a short introduction.

* Update Python 3 API reference

An API section has been added. Only the Python API is present and later
the Qbs API will be added.

Also note there are several improvements on the directives.  Finally,
the global.rst file is removed since there is no real use for now since
the glossary is already in the master-toc.

* Update documentation from the stoiridh.qbs.tools.qbs module

Minor changes in order to reflect those done within the
stoiridh.qbs.tools module.