Add a Sphinx extension in order to support Qbs
The extension is in an alpha state 0.1.0-alpha and is based on the
Python domain.

With the Qbs domain, we are able to document both Qbs items and Qbs
modules. Since Qbs modules have a different meaning compared to Python,
a 'sdk' directive has been setting up which fulfils the same role like
the module directive from the Python domain.

Both Qbs items and Qbs modules support the 'property' directive in order
to document their properties. The property signature is the following:
type, name[, value]. A property may specify the 'readonly' option.

Finally, the Sphinx extension is licenced under the BSD licence which is
compatible with the GPL licence.

Task-number: #12