Add PyPI support (#21)
* Rename setup.py

This file has been renamed to stoiridhtools.py

* Update .gitignore

In order to integrate with PyPI, we must ignore the 'build' directory
generated in the top-level project.

* Rename stoiridh package

The package has been renamed to the project's name.

* Reogarnise the package-tree

Including the update of the test files.

* Move the share directory

The directory is moved into the stoiridhtools package in order to be
install through pip.

* Move stoiridhtools.py

This file has been renamed to cli.py and move into the stoiridhtools package.
A bootstrap script for the stoiridhtools.cli module (stoiridh-cli.py) has
been added. It is be used as a wrapper for the module.

* Replace README file format

In favour to the reStructuredText format.

* Update documentation

This update reflects the changes bring to the stoiridhtools package.

* Add PyPI support

This will allow us to install the package through pip.