Update the stoiridhtools.cli module (#23)
* Rename sqt.conf files

With the reboot of the project, sqt (Stòiridh Qbs Tools) doesn't make
sens anymore. So it has been renamed by its full name.

* Improve stoiridhtools.Config class

The class comes with a new classmethod in order to get the default path.
The path parameter from the __init__() method becomes optional and if it
is not specified, then use the default path.

* Improved stoiridhtools.SDK class

The __init__() method comes with a new optional parameter, path. This
parameter will allow the install of the packages from the SDK elsewhere
else if need. Of course, if the path parameter is not specified, then it
will use the default path from the stoiridhtools.Config class.

* Improve stoiridhtools.cli module

Add the possibility to add several command that will be handled by the
stoiridhtools.cli.CommandManager class which is a context manager.