Improve verbosity
Instead of to restrict the verbosity of the messages to the
stoiridhtools.cli.Command class, move the verbosity handling directly
into the stoiridhtools module in order that all classes from the modules
can be able to print verbose messages.

Note also that the constants written in the stoiridhtools.cli submodule
has been moved and improved into the stoiridhtools module.