Add a configuration command for the cli
The stoiridhtools.cli.config.ConfigCommand class will allow to change
the Stòiridh Tools settings. Only Qt is implemented as a subcommand for
the config command for now.

The stoiridhtools.cli.config.QtConfigCommand class is a subcommand for
the ConfigCommand class. This class has one option which allows to set
the HOME path where Qt is installed.