Add a configuration command for the cli (#29)
* Add subcommands support

A Command object now can have one or more Command objects as
subcommands.

CommandManager has been improved and optimised in order to handle those
subcommands.

The InitCommand class has been updated in order to use the new
functionality to create a command.

* Add get_description() method

This method set in the Command class allows subclasses to provide a
brief description of their command in order to be printed when the user
will display the help of the command.

* Improve arguments handling

Instead to get an error when there is a unknown argument, let this case
be handled by the commands by setting up a parameter named unknown_args
that other commands can access if desired, in their run method.

* Add a configuration command for the cli

The stoiridhtools.cli.config.ConfigCommand class will allow to change
the Stòiridh Tools settings. Only Qt is implemented as a subcommand for
the config command for now.

The stoiridhtools.cli.config.QtConfigCommand class is a subcommand for
the ConfigCommand class. This class has one option which allows to set
the HOME path where Qt is installed.