Add support for reading a Qbs configuration file
We are now able to extract the profiles held in a Qbs configuration
file.

As for the Profile class itself, it contains a limited access to these
data for now. We can get the name, the Qbs settings and the preferences
of the profile as well as add new Qbs search paths.