Fix an issue in the _make_value function
This issue occurred only on GNU/Linux when we wanted to check the
validity of the value parameter in order to know whether that is a path.
The problem here was we wanted to check the validity of a path on a
list, which is wrong. So, we must move this condition at the end of the
if statement. When the value parameter will be a list, then it will meet
the first condition in which it will transform into a list.