Add support for reading a Qbs configuration file (#33)
* Add support for reading a Qbs configuration file

We are now able to extract the profiles held in a Qbs configuration
file.

As for the Profile class itself, it contains a limited access to these
data for now. We can get the name, the Qbs settings and the preferences
of the profile as well as add new Qbs search paths.

* Fix an issue about the path separators

This issue figures out a mistake about the path separators which don't
automatically seem to be updated according to the OS.


* Fix an issue in the _make_value function

This issue occurred only on GNU/Linux when we wanted to check the
validity of the value parameter in order to know whether that is a path.
The problem here was we wanted to check the validity of a path on a
list, which is wrong. So, we must move this condition at the end of the
if statement. When the value parameter will be a list, then it will meet
the first condition in which it will transform into a list.

Task-number: #33