Add the logging API
The logging API is based on the standard module 'logging'. Here we use
the Proxy design pattern in order to modify the default behaviour by
colouring and customising the output.

Note: The entry point of the program has been changed since the adding
of this feature. Instead to load the main function from the CLI
submodule, we load the main function of the module itself.

Task-number: #38