Add the logging API (#45)
* Add the logging API

The logging API is based on the standard module 'logging'. Here we use
the Proxy design pattern in order to modify the default behaviour by
colouring and customising the output.

Note: The entry point of the program has been changed since the adding
of this feature. Instead to load the main function from the CLI
submodule, we load the main function of the module itself.

* Add requirements.txt

This file will install the dependencies of the projects like colorama
for the Stòiridh Tools' logging system.

* Update .gitignore

This update allows to ignore the .python-version file required by PyEnv.

* Sort the imports

Reorganise the way of importing the modules.

* Update CI configuration file

Install requirements.txt in order to be able to test the project.

Task-number: #38